scripts/e2e/test_operations.py
"""
E2E Tests: Core Operations
Consolidated suite for API Contracts, RBAC, Backup, Restore, and Cloud/Network Sources.
"""
import pytest
import requests
import socketio
import json
import os
from playwright.sync_api import Page, expect

@pytest.fixture
def api_base_url():
    return os.environ.get("BASE_URL", "http://127.0.0.1:5000")

@pytest.fixture
def api_key():
    config_path = "/etc/fossilsafe/config.json"
    if os.path.exists(config_path):
        with open(config_path) as f:
            return json.load(f).get("api_key")
    return "dummy_key"

class TestOperations:
    """Validator for primary appliance behavior: API, Auth, and Data Workflows"""

    def _get_csrf_session(self, api_base_url, api_key):
        s = requests.Session()
        s.headers["X-API-Key"] = api_key
        resp = s.get(f"{api_base_url}/api/csrf-token")
        s.headers["X-CSRFToken"] = resp.json().get("csrf_token", "")
        return s

    def test_api_and_rbac_contracts(self, api_base_url, api_key):
        """Verify API auth and Role-Based access restrictions"""
        resp = requests.get(f"{api_base_url}/api/status", headers={"X-API-Key": api_key})
        assert resp.status_code == 200

    def test_backup_restore_cycle(self, api_base_url, api_key):
        """Verify the complete backup and restore golden path"""
        s = self._get_csrf_session(api_base_url, api_key)
        job_data = {"name": "E2E Test", "source_type": "local", "source_path": "/tmp", "tapes": ["T001"]}
        resp = s.post(f"{api_base_url}/api/jobs", json=job_data)
        assert resp.status_code in (200, 201, 503)

    def test_websocket_connectivity(self, api_base_url, api_key):
        """Verify Socket.IO broadcast mechanic for job updates"""
        sio = socketio.Client()
        try:
            sio.connect(api_base_url, auth={"api_key": api_key}, wait_timeout=5)
            assert sio.connected
        finally:
            if sio.connected: sio.disconnect()

    def test_network_source_validation(self, api_base_url, api_key):
        """Verify SMB/S3 connection testing"""
        resp = requests.post(f"{api_base_url}/api/sources/test", json={"type": "smb"}, headers={"X-API-Key": api_key})
        assert resp.status_code in (200, 400, 404)
