requirements.txt
APScheduler==3.10.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==42.0.8
dnspython==2.8.0
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
gevent==25.9.1
greenlet==3.3.1
gunicorn==21.2.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
ply==3.11
pyasn1==0.6.2
pycparser==3.0
pycryptodomex==3.23.0
Pygments==2.19.2
pyotp==2.9.0
pysmi==1.6.3
pysnmp==4.4.12
pytest==9.0.2
pytest-asyncio==1.3.0
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
tzlocal==5.3.1
urllib3==2.6.3
websocket-client==1.8.0
Werkzeug==3.1.5
wsproto==1.3.2
WTForms==3.2.1
zope.event==6.1
zope.interface==8.2
boto3==1.34.40
rich==13.7.1
